refact: make UITextEditor more reusable
